Why UA sniffing is generally bad, real-world example #1048577: "That time a missing /s on navigator.userAgent (based on single-digit UA version assumptions) broke SM2 on iOS 8+ devices."
The upside: The fix is adding \s to a UA regex check targeting legacy iOS devices with broken HTML5 Audio support, so that the "OS 10 " portion of the UA string does not match the pattern "OS 1 ".

This is a pre-emptive fix for the pending iOS 8 release, which reportedly includes the string "iPhone OS 10_10 like Mac OS X". Due to SM2 checking for /os ([12|3_0|3_1])/i which had broken HTML5 Audio() support, and not considering the far future possibility of "iOS 10" being included in the userAgent string at the time, HTML5 audio is incorrectly disabled on iOS 8 beta 3 and is likely to apply to the final iOS 8 release.

This is a good example and reminder of why UA checks are generally bad, and dangerous; similar issues hit other libraries doing UA and plugin checks both client and server-side when IE hit 10, and Flash hit version 10, so this iOS pattern change may affect some other libraries targeting iOS as well.