Reverted from CustomEvent to standard Event
Reverted from CustomEvent to standard Event for IE compatibility as
well as simplicity.  There was a bug in the properties I passed as
well, but I’m not going to get into it and I wasn’t able to properly
pass custom event details for some reason anyways. This works just fine.