Feature: Allow custom SWF url
Currently, it is only possible to set the
url of the directory where the SWF files
exist, not the exact filename.

This PR adds an extra option, 'customSwfUrl',
which, if set, will override the logic for
determining path to a SWF file, to use the
custom SWF url directly.

This makes it possible to use SoundManager
with other filenames than the default ones,
e.g. with hashes added. It also makes it
easier to use SoundManager with Webpack
(or another module bundler):

```javascript
soundManager.setup({
  customSwfUrl: require('soundmanager2/swf/soundmanager2_flash9.swf'),
  /// ...
});
```